cmd/kcp-front-proxy move filters into module
